roar_py_core/requirements.txt
gymnasium==0.28.1
numpy
asyncio
pyserde[all]==0.10.4
Pillow==9.5.0
transforms3d==0.4.1
pytest>=7.3.1
pytest-asyncio>=0.21.0
networkx==3.1
matplotlib
websockets>=11.0.1
typing_inspect>=0.8.0